+++ packages/hooks/src/register-extension.js
@@ -12,7 +12,7 @@
 
 const { appendAction } = require('./state');
 const { logAction } = require('./logger');
-const { getHook } = require('./create-actions-registry');
+const { getAction } = require('./create-actions-registry');
 
 const registerExtension = (__arg1 = {}, __arg2 = null, __arg3 = {}) => {
   // (name, handler, options)
@@ -77,7 +77,7 @@ const registerExtension = (__arg1 = {}, __arg2 = null, __arg3 = {}) => {
   try {
     const hook =
       targetAction.substr(0, 1) === '$'
-        ? getHook(targetAction.substr(1))
+        ? getAction(targetAction.substr(1))
         : targetAction;
 
     // the hook could be null in case of an optional reference was required